arreglados los nombres de las cajas

+++ server/routes/virtualCashPointsRoutes.js
@@ -20,15 +20,32 @@ router.get("/:idcashPoint", async (req, res) => {
 router.post("/", async (req, res) => {
   const { name, idCashPoint } = req.body;
   try {
-    // Reiniciar la secuencia para la columna idGlobalVirtualCashPoint al valor más alto utilizado previamente
-    await db.one(
-      "SELECT setval('virtualcashpoint_idglobalvirtualcashpoint_seq', COALESCE((SELECT MAX(idGlobalVirtualCashPoint) + 1 FROM VirtualCashPoint), 1), false);"
+    // Obtener el valor máximo de idVirtualCashPoint para el idCashPoint actual
+    const maxIdVirtualCashPointResult = await db.oneOrNone(
+      "SELECT maxIdVirtualCashPoint FROM MaxVirtualCashPointSeq WHERE idCashPoint = $1",
+      [idCashPoint]
     );
 
+    let nextIdVirtualCashPoint = 1;
+    if (maxIdVirtualCashPointResult) {
+      nextIdVirtualCashPoint =
+        maxIdVirtualCashPointResult.maxidvirtualcashpoint + 1;
+    }
+
+    console.log("nextIdVirtualCashPoint: " + nextIdVirtualCashPoint);
+
+    // Insertar el nuevo cajero virtual
     const newVirtualCashPoint = await db.one(
-      "INSERT INTO VirtualCashPoint (name, idCashPoint) VALUES ($1, $2) RETURNING *",
-      [name, idCashPoint]
+      "INSERT INTO VirtualCashPoint (idVirtualCashPoint, name, idCashPoint) VALUES ($1, $2, $3) RETURNING *",
+      [nextIdVirtualCashPoint, name, idCashPoint]
     );
+
+    // Actualizar o insertar el valor máximo de idVirtualCashPoint en MaxVirtualCashPointSeq
+    await db.none(
+      "INSERT INTO MaxVirtualCashPointSeq (idCashPoint, maxIdVirtualCashPoint) VALUES ($1, $2) ON CONFLICT (idCashPoint) DO UPDATE SET maxIdVirtualCashPoint = $2",
+      [idCashPoint, nextIdVirtualCashPoint]
+    );
+
     res.json(newVirtualCashPoint);
   } catch (error) {
     res.status(500).json({ error: error.message });
@@ -54,7 +71,7 @@ router.put("/:id", async (req, res) => {
 router.delete("/:id", async (req, res) => {
   const id = req.params.id;
   try {
-    // Si se esta eliminando una caja que tiene usuarios asignados, muestra un mensaje de error ya que deberia eliminar primero los usuarios asignados a esa caja
+    // Si se está eliminando un cajero que tiene usuarios asignados, muestra un mensaje de error ya que debería eliminar primero los usuarios asignados a ese cajero
     const users = await db.any(
       'SELECT * FROM "User" where idGlobalVirtualCashPoint=$1',
       [id]
@@ -68,12 +85,39 @@ router.delete("/:id", async (req, res) => {
       return;
     }
 
-    await db.none(
-      "DELETE FROM VirtualCashPoint WHERE idGlobalVirtualCashPoint=$1",
+    // Antes de eliminar, obtener el idCashPoint asociado con este cajero virtual
+    const idCashPointResult = await db.oneOrNone(
+      "SELECT idCashPoint FROM VirtualCashPoint WHERE idGlobalVirtualCashPoint=$1",
       [id]
     );
 
-    res.json({ message: "Cajero virtual eliminado exitosamente" });
+    if (idCashPointResult) {
+      const idCashPoint = idCashPointResult.idCashPoint;
+
+      // Eliminar el cajero virtual
+      await db.none(
+        "DELETE FROM VirtualCashPoint WHERE idGlobalVirtualCashPoint=$1",
+        [id]
+      );
+
+      // Actualizar o eliminar el registro de MaxVirtualCashPointSeq según sea necesario
+      const remainingVirtualCashPoints = await db.oneOrNone(
+        "SELECT COUNT(*) FROM VirtualCashPoint WHERE idCashPoint = $1",
+        [idCashPoint]
+      );
+
+      if (remainingVirtualCashPoints === 0) {
+        // Si ya no quedan cajeros virtuales para este idCashPoint, eliminar el registro de MaxVirtualCashPointSeq
+        await db.none(
+          "DELETE FROM MaxVirtualCashPointSeq WHERE idCashPoint = $1",
+          [idCashPoint]
+        );
+      }
+
+      res.json({ message: "Cajero virtual eliminado exitosamente" });
+    } else {
+      res.status(404).json({ message: "Cajero virtual no encontrado" });
+    }
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
